+++ src/model/graphql-response.ts
@@ -1,0 +1,16 @@
+export interface GraphqlResponse {
+	/** An array that contains any GraphQL errors. */
+	errors?: GraphqlError[];
+	/** If the data was loaded without any errors, contains the requested object. */
+	data?: object;
+}
+
+interface GraphqlError {
+	message: string;
+	errorType: string;
+	validationErrorType?: string;
+	queryPath: string[];
+	path?: any;
+	extensions?: any;
+	locations: { line: number; column: number; sourceName: string }[];
+}
