+++ src/services/webhooks/webhooks.service.ts
@@ -1,4 +1,4 @@
-import ApiBase, { SuccessResponse } from '../../api-base';
+import ApiBase, { SuccessResponse } from '../api-base/api-base';
 import Webhook from '../../model/webhook';
 import { extractId, IdOrObject } from '../../util/id-or-object';
